docker/sagemaker-endpoint/requirements.txt
torch==2.2.1
torchvision==0.17.1
sentence-transformers==2.6.1
faiss-cpu==1.8.0
fastapi==0.110.0
uvicorn==0.29.0